"Move 'upgrade plan' functionality to 'upgrade plan cluster'
This is more in sync with our 'verb + resource' pattern for commands
plus it opens the possibility to write upgrade plan commands for other
types of resources.
